WindowManager class is added, made changes to Game
WindowManager class is added and made some changes to the Game class.
WindowManager's purpose is to have Window Objects contained in a vector
and managed by the class. Since the Window class is a derivative it's a
NonCopyable. Therefore WindowManager's Window container will contain
pointers to the Window Class objects.