Just added a comment line
Added a comment line for returning to project where I left. I probably
will be implementing some file operations classes in order to read
settings from files. Next step will be creating some file or method or
something that keeps default settings. And if there is not any setting
file, or even there is and the searched settings can't be found, the
default ones will be used.